component-tests/requirements.txt
cloudflare==2.8.15
flaky==3.7.0
pytest==6.2.2
pyyaml==5.4.1
requests==2.25.1
retrying==1.3.3